Use a block to scope the mapping to a specific app
When you use an <item>, it produces a checkbox for you to enable/disable
the mapping in the GUI. I don't need or want that for this particular
mapping. I just want a way to define a one mapping (among a set of
mappings) where the the mapping is scoped to a specific app. A <block>
meets that need.