Enable Command+Option+Up/Down in (S)uper (D)uper mode
In (S)uper (D)uper mode, 'A' is Option, 'F' is Command, 'J' is Up, and 'K' is
Down. It's useful to be able to press Command+Option+Up and Command+Option+Down
(e.g., to switch panes in iTerm 2).

Prior to this commit, when you tried to Command+Option+Up/Down in (S)uper
(D)uper mode, the rules for Option+Up/Down kicked in, and it resulted in Page
Up/Page Down. In other words, it ignored the fact that Command was pressed. This
commit ensures that the Page Up/Page Down mapping kicks in _only_ for
Option+Up/Down. This allows A+F+J/K to trigger Command+Option+Up/Down in (S)uper
(D)uper mode.