Enable Command+Option+Left/Right in SD mode for iTerm 2
In (S)uper (D)uper mode, 'A' is Option, 'F' is Command, 'H' is Left, and 'L' is
Right. In iTerm 2, it's useful to be able to press Command+Option+Left and
Command+Option+Right to switch panes.

Prior to this commit, when you tried to Command+Option+Left/Right in (S)uper
(D)uper mode in iTerm 2, the rules for Option+Left/Right kicked in, and it
resulted in moving to the previous/next word. In other words, it ignored the
fact that Command was pressed. This commit ensures that the previous/next word
mapping kicks in _only_ for Option+Left/Right. This allows A+F+H/L to trigger
Command+Option+Left/Right in (S)uper (D)uper mode for iTerm 2.