Fix Markdown Mode bug in Chrome
For some reason, the old Markdown Mode code recently stopped working in
Chrome. I think this has to due with how `osascript` and AppleScript
work. As far as I know, AppleScript is asynchronous. So, the `osascript`
commands kick off the AppleScript keystrokes, and then the `osascript`
command exits. The keystrokes may not yet have completed when the
`osascript` command exits.

If we could force the AppleScript commands to run synchronously, that
would resolve the issue, but I'm not aware of a way to do that.

We can introduce a sleep statement to increase the likelihood that the
keystrokes have completed before we run the remaining commands. This is
far from elegant, but I'm not sure of another solution at the moment.