Define keybinding earlier
When working on these configs, if there's an error in the code, it
will prevent the subsequent code from getting executed. In those cases,
we need to fix the issue and reload the config. To make it easier to
reload the config, it's helpful to bind the hotkey for reloading the
config *before* any of the other config code.