Construct status message drawings lazily
Hammerspoon doesn't appear to have a way to indicate which screen should
receive the drawing. It seems to always go to the primary screen. If you
connect or disconnect a monitor while Hammerspoon is running, your
primary screen may change. Since we use the dimensions of the primary
screen to calculate the coordinates for the status message, and since
the coordinates of the primary screen can change, we need to dynamically
determine the coordinates of the status message each time we display it
(or, we need to recalculate the coordinates any time the primary screen
changes). For now, the easiest approach seems to be to lazily construct
the status message drawings each time we need to display them.