Symlink karabiner dir instead of just config file
Due to limitation in mac osx, the file alone isn't enough to trigger changes when the file updates. This meant that I needed to restart karabiner-elements after switching profiles or any changes to a profile. Moving to the dir fixed the issue.

xref https://github.com/tekezo/Karabiner-Elements/issues/597#issuecomment-282760186