Resolve controlKeyTimer name collision
I introduced this bug in c0e915b when I renamed control_key_timer to
controlKeyTimer. It turns out `microphone.lua` also has a function of
the same name.

To address this, I saw two potential solutions. First, renaming the
function to alleviate the conflict altogether. And second, declaring
functions as `local`. I was tempted to implement the latter, and
successfully did so locally, but decided against submitting that work as
I'm not fully confident in my understanding of lexical scoping in Lua.