Configure Bear to paste as plaintext by default
Without this change, whenever you use command+v to paste a URL in Bear, 
Bear formats the URL as Markdown. For example, if you paste 
http://example.com, Bear converts it to:

[http://example.com](http://example.com)

That might be what some people want, but that's just not my jam.

With this change, command+v will always paste as plaintext, thus 
avoiding the behavior seen above. 😅