Remove push-to-talk/push-to-mute functionality
Fixes #77

To keep this repository focused on things that I personally find useful
and enjoyable to maintain [1], this commit removes the push-to-talk and
push-to-mute functionality, since I don't actually use it. I'm not sure
why it never clicked with my workflow, because I certainly think it's
cool.

This also has the benefit of fixing #77. It seems like something changed
recently in Hammerspoon that causes the status message from
hammerspoon/microphone.lua to persist on startup. Since this commit
removes hammerspoon/microphone.lua, it also resolves this bug.

[1] https://github.com/jasonrudolph/keyboard/tree/ddffacf#more-useful-for-me